refactor login routes
