+++ src/app.js
@@ -1,88 +1,17 @@
 'use strict';
 
 require('dotenv').config();
-
 const express = require('express');
-const passport = require('passport');
-const Strategy = require('passport-twitter').Strategy;
 
-let trustProxy = false;
-if (process.env.DYNO) {
-  trustProxy = true;
-}
-
-passport.use(
-  new Strategy(
-    {
-      consumerKey: process.env.TWITTER_CONSUMER_KEY,
-      consumerSecret: process.env.TWITTER_CONSUMER_SECRET,
-      callbackURL: '/oauth/callback',
-      proxy: trustProxy,
-    },
-    function(token, tokenSecret, profile, cb) {
-      return cb(null, profile);
-    }
-  )
-);
-
-passport.serializeUser(function(user, cb) {
-  cb(null, user);
-});
-
-passport.deserializeUser(function(obj, cb) {
-  cb(null, obj);
-});
+const loginRouter = require('./login');
 
 const app = express();
 app.use(express.static('public'));
 app.use(require('morgan')('combined'));
 app.use(require('body-parser').urlencoded({ extended: true }));
-// TODO: change the secret and move to .env
-// Originally the resave and saveUninitialized were set to true, changing to false didn't seem to affect anything. Look into the docs for how this works
-app.use(require('express-session')({ secret: 'keyboard cat', resave: false, saveUninitialized: false }));
+app.use(require('express-session')({ secret: process.env.SESSION_SECRET, resave: false, saveUninitialized: false }));
 
-app.use(passport.initialize());
-// TODO: maybe delete
-// app.use(passport.session());
-
-app.get('/login/twitter', passport.authenticate('twitter'), (request, response) => {
-  // this response should never be displayed to user if everything works correctly
-  response.send({ hello: 'this is the auth route' });
-});
-
-app.get(
-  '/oauth/callback',
-  passport.authenticate('twitter', { failureRedirect: '/login/twitter' }),
-  (request, response) => {
-    const sessionId = request.sessionID;
-    const sessionData = request.sessionStore.sessions[sessionId];
-    const oAuthData = JSON.parse(sessionData)['oauth:twitter'];
-
-    const userData = request.user._json;
-
-    const savedUserData = {
-      userId: userData.id,
-      userName: userData.name,
-      userScreenName: userData.screen_name,
-      photoLink: request.user.photos[0].value,
-      oAuthToken: request.query.oauth_token,
-      oAuthVerifier: request.query.oauth_verifier,
-      oAuthTokenSecret: oAuthData.oauth_token_secret,
-    };
-
-    // TODO: add to database
-
-    response.redirect(
-      `exp://exp.host/@melissastock/front-end/?display_name=${savedUserData.userScreenName}&user_name=${
-        savedUserData.userName
-      }&id=${savedUserData.userId}`
-    );
-  }
-);
-
-app.get('/dashboard', (request, response) => {
-  response.send('Logged in!');
-});
+app.use('/', loginRouter);
 
 app.get('/logout', (request, response) => {
   request.session.destroy((err) => {
